+++ common/src/main/java/com/st0x0ef/stellaris/client/overlays/RocketBarOverlay.java
@@ -3,15 +3,18 @@ package com.st0x0ef.stellaris.client.overlays;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.st0x0ef.stellaris.Stellaris;
 import com.st0x0ef.stellaris.client.screens.helper.ScreenHelper;
+import com.st0x0ef.stellaris.common.entities.vehicles.LanderEntity;
 import com.st0x0ef.stellaris.common.entities.vehicles.RocketEntity;
 import com.st0x0ef.stellaris.common.utils.PlanetUtil;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.RenderType;
+import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Blocks;
 
 public class RocketBarOverlay {
 
@@ -20,29 +23,30 @@ public class RocketBarOverlay {
     public static void render(GuiGraphics graphics, DeltaTracker deltaTracker) {
         Player player = Minecraft.getInstance().player;
 
-        if (player.getVehicle() instanceof RocketEntity
-                //|| player.getVehicle() instanceof LanderEntity
-                ) {
+        if (player.getVehicle() instanceof RocketEntity || player.getVehicle() instanceof LanderEntity) {
             Level level = Minecraft.getInstance().level;
 
-            float yHeight = (float) player.getY() / 5.3F;
+            if (level == null) return;
 
-            if (yHeight < 0) {
-                yHeight = 0;
-            } else if (yHeight > 113) {
-                yHeight = 113;
+            double min = player.level().getMinBuildHeight();
+
+            for (int i = player.level().getMinBuildHeight(); i < player.level().getMaxBuildHeight(); i++) {
+                if (level.getBlockState(new BlockPos(player.getBlockX(), i, player.getBlockZ())).getBlock() != Blocks.AIR) {
+                    min = i;
+                }
             }
+
+            double yHeight = ((player.getY() - min) / (600 - min)) * 113;
 
             ResourceLocation planet = PlanetUtil.getPlanetBar(level.dimension().location());
 
             /** ROCKET BAR IMAGE */
             RenderSystem.setShaderTexture(0, planet);
-            graphics.blit(RenderType::guiTextured, planet, 0, (graphics.guiHeight() / 2) - 128 / 2, 0, 0, 16, 128, 16, 128);
+            graphics.blit(planet, 0, (graphics.guiHeight() / 2) - 128 / 2, 0, 0, 16, 128, 16, 128);
 
             /** ROCKET_Y IMAGE */
             RenderSystem.setShaderTexture(0, ROCKET);
-            ScreenHelper.renderWithFloat.blit(graphics.pose(), 4, ((float) graphics.guiHeight() / 2) + ((float) 103 / 2) - yHeight, 0, 0, 8, 11, 8, 11);
+            graphicst.blit(RenderType::guiTextured, graphics.pose(), 4F, (float) (((float) graphics.guiHeight() / 2) + (103f / 2f) - yHeight), 0, 0, 8, 11, 8, 11);
         }
-
     }
 }