+++ common/src/main/java/com/st0x0ef/stellaris/common/blocks/entities/machines/CoalGeneratorEntity.java
@@ -112,7 +112,7 @@ public class CoalGeneratorEntity extends BaseGeneratorBlockEntity {
             }
         }
 
-        EnergyApi.distributeEnergyNearby(this,100);
+        EnergyApi.distributeEnergyNearby(this, getMaxCapacity());
     }
 
     protected int getBurnDuration(ItemStack fuelStack) {