+++ common/src/main/java/com/st0x0ef/stellaris/common/blocks/entities/machines/FuelRefineryBlockEntity.java
@@ -64,18 +64,22 @@ public class FuelRefineryBlockEntity extends BaseEnergyContainerBlockEntity impl
         if (this.level.isClientSide) return;
 
         if (getItem(2).getItem() instanceof JetSuit.Suit) {
-            int fuel = 1000;
+            int fuel = FluidTankHelper.convertFromNeoMb(10);
 
-            if(resultTank.getAmount() < 1000) {
+            if (resultTank.getAmount() < fuel) {
                 fuel = (int) resultTank.getAmount();
             }
-            if (JetSuit.Suit.getFuel(getItem(2)) + fuel > 243_000) {
-                fuel = 243000 -  (int)JetSuit.Suit.getFuel(getItem(2));
+
+            else if (FuelUtils.getFuel(getItem(2)) + fuel > JetSuit.MAX_FUEL_CAPACITY) {
+                fuel = (int) (JetSuit.MAX_FUEL_CAPACITY - (int) FuelUtils.getFuel(getItem(2)));
             }
 
-            JetSuit.Suit.addFuel(getItem(2), fuel);
-            resultTank.grow(-fuel);
-
+            if (FuelUtils.addFuel(getItem(2), fuel)) {
+                resultTank.shrink(fuel);
+                this.setChanged();
+            }
+        } else {
+            FluidTankHelper.extractFluidToItem(this, resultTank, 2, 3);
         }
     }