+++ common/src/main/java/com/st0x0ef/stellaris/common/blocks/entities/machines/WaterSeparatorBlockEntity.java
@@ -97,8 +97,6 @@ public class WaterSeparatorBlockEntity extends BaseEnergyContainerBlockEntity im
 
         if (!resultTanks.getLast().isEmpty()) {
             FluidTankHelper.transferFluidNearby(this, resultTanks.getLast());
-
-
         }
     }