+++ common/src/main/java/com/st0x0ef/stellaris/common/entities/vehicles/base/AbstractRoverBase.java
@@ -345,8 +345,7 @@ public abstract class AbstractRoverBase extends IVehicleEntity
     }
 
     public boolean isAccelerating() {
-        boolean b = (isForward() || isBackward()) && !horizontalCollision;
-        return b;
+        return (isForward() || isBackward()) && !horizontalCollision;
     }
 
     @Override